requirements.txt
googleapiclient
google
imutils
flask
Flask-BasicAuth==0.2.0
requests
python-telegram-bot
tzlocal
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib